Add size property to button
