+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/semanticui/elements/button/Button.scala
@@ -1,5 +1,6 @@
 package edu.gemini.seqexec.web.client.semanticui.elements.button
 
+import edu.gemini.seqexec.web.client.semanticui.Size
 import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.vdom.prefix_<^._
@@ -24,6 +25,7 @@ object Button {
                    emphasis: Emphasis    = NoEmphasis,
                    animated: Animated    = NotAnimated,
                    icon: Option[Icon]    = None,
+                   size: Size            = Size.NotSized,
                    basic: Boolean        = false,
                    inverted: Boolean     = false,
                    circular: Boolean     = false,
@@ -46,7 +48,14 @@ object Button {
       "inverted"  -> p.inverted,
       "circular"  -> p.circular,
       "labeled"   -> p.labeled,
-      "disabled"  -> p.disabled
+      "disabled"  -> p.disabled,
+      "tiny"      -> (p.size == Size.Tiny),
+      "mini"      -> (p.size == Size.Mini),
+      "small"     -> (p.size == Size.Small),
+      "large"     -> (p.size == Size.Large),
+      "big"       -> (p.size == Size.Big),
+      "huge"      -> (p.size == Size.Huge),
+      "massive"   -> (p.size == Size.Massive)
     )
 
   def component = ReactComponentB[Props]("Button")
